requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
ccxt==4.1.100
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
numpy==1.26.2
pandas==2.1.4
pycares==4.4.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
